+++ packages/terra-compact-interactive-list/src/proptypes/columnShape.js
@@ -10,7 +10,7 @@ const columnShape = PropTypes.shape({
   /**
    * String of text which will be used to provide a column context and to serve as a column header if needed.
    */
-  displayName: PropTypes.string,
+  displayName: PropTypes.string.isRequired,
 
   /**
    * A valid css string, px, em, or rem supported (should be the same across all width units).