+++ packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples/ResponsiveColumns.jsx
@@ -48,19 +48,19 @@ const rows = [
 const cols = [
   {
     id: 'Column-0',
-    displayName: 'Col_1',
+    displayName: 'Icon',
     width: '60px',
     align: alignTypes.CENTER,
   },
   {
     id: 'Column-1',
-    displayName: 'Col_2',
+    displayName: 'Service name',
     width: '200px', // will be disregarded because of the flexGrow prop
     flexGrow: true, // flexGrow prop has to be set here, as with width set, without that prop the column would not grow.
   },
   {
     id: 'Column-2',
-    displayName: 'Col_3',
+    displayName: 'Featured',
     width: '60px',
     align: alignTypes.CENTER,
   },
@@ -72,6 +72,7 @@ const ResponsiveColumns = () => (
     rows={rows}
     columns={cols}
     numberOfColumns={2}
+    rowHeaderIndex={1}
   />
 );