+++ packages/terra-framework-docs/src/terra-dev-site/doc/theme-provider/example/DefaultThemeProvider.jsx
@@ -11,16 +11,23 @@ class DefaultThemeProvider extends React.Component {
     super(props);
     this.state = {
       theme: '',
+      density: '',
     };
     this.handleThemeChange = this.handleThemeChange.bind(this);
+    this.handleDensityChange = this.handleDensityChange.bind(this);
   }
 
-  handleThemeChange(e) {
-    this.setState({ theme: e.target.value });
+  handleThemeChange(event) {
+    this.setState({ theme: event.target.value });
+  }
+
+  handleDensityChange(event) {
+    this.setState({ density: event.target.value });
   }
 
   render() {
     let themeSwitcher;
+    let densitySwitcher;
 
     function supportsCSSVars() {
       // This eslint rule complains of the lack of .CSS support in IE, which is exactly what this line is intended to detect.
@@ -38,6 +45,18 @@ class DefaultThemeProvider extends React.Component {
           </select>
         </div>
       );
+
+      densitySwitcher = (
+        <div className={cx('theme-switcher-wrapper')}>
+          <label htmlFor="density"> Density: </label>
+          <select id="density" value={this.state.density} onChange={this.handleDensityChange}>
+            <option value="">Default</option>
+            <option value="compact">Compact</option>
+            <option value="comfortable">Comfortable</option>
+            <option value="spacious">Spacious</option>
+          </select>
+        </div>
+      );
     } else {
       themeSwitcher = (
         <div />
@@ -47,7 +66,8 @@ class DefaultThemeProvider extends React.Component {
     return (
       <div>
         <div>{themeSwitcher}</div>
-        <ThemeProvider themeName={this.state.theme}>
+        <div>{densitySwitcher}</div>
+        <ThemeProvider themeName={this.state.theme} density={this.state.density}>
           <div>
             <MockThemeComponent>Themable component</MockThemeComponent>
           </div>