+++ packages/terra-framework-docs/src/terra-dev-site/test/compact-interactive-list/FixedWidthColumns.test.jsx
@@ -50,18 +50,18 @@ const rows = [
 const cols = [
   {
     id: 'Column-0',
-    displayName: 'Col_1',
+    displayName: 'Icon',
     width: '60px',
     align: alignTypes.CENTER,
   },
   {
     id: 'Column-1',
-    displayName: 'Col_2',
+    displayName: 'Service name',
     width: '350px',
   },
   {
     id: 'Column-2',
-    displayName: 'Col_3',
+    displayName: 'Featured',
     width: '60px',
     align: alignTypes.CENTER,
   },
@@ -75,6 +75,7 @@ const FixedWidthColumns = () => (
     numberOfColumns={2}
     columnMinimumWidth="20px" // should have no effect of fixed columns
     columnMaximumWidth="200px" // should have no effect of fixed columns
+    rowHeaderIndex={1}
   />
 );