+++ packages/terra-pills/src/FilterPills.jsx
@@ -94,7 +94,6 @@ const FilterPills = (props) => {
   const isPillDeleted = useRef(false);
   const isRollUpRemoved = useRef(false);
   const removedLabel = useRef();
-  const containerHint = useRef(ariaLabel);
 
   // Identifies the number of pills that needs to be hidden/rolled up
   const generateRollUp = useCallback(() => {
@@ -286,7 +285,8 @@ const FilterPills = (props) => {
         event.preventDefault();
         focusPreviousNode(pills, rollUpPill);
         break;
-      case KEY_BACK_SPACE || KEY_DELETE:
+      case KEY_BACK_SPACE:
+      case KEY_DELETE:
         event.preventDefault();
         focusNodeAfterDelete(pills);
         break;
@@ -378,7 +378,7 @@ const FilterPills = (props) => {
   if (isCollapsible && (rollUpCount > 0) && isCollapsed) {
     pillGroupInteractionHint += `, ${intl.formatMessage({ id: 'Terra.pills.hint.rollupNotVisible' }, { pillsNotVisibleCount: rollUpCount })}`;
   } else if (isCollapsible && !isCollapsed && showRollupPillInteraction) {
-    removedPillInteractionHint = intl.formatMessage({ id: 'Terra.pills.hint.rollupVisible' });
+    pillGroupInteractionHint = intl.formatMessage({ id: 'Terra.pills.hint.rollupVisible' });
   }
   if (pillRemoved) {
     removedPillInteractionHint = intl.formatMessage({ id: 'Terra.pills.hint.wasRemoved' }, { pillLabelName: removedLabel.current });
@@ -397,24 +397,28 @@ const FilterPills = (props) => {
     return reducedArray;
   };
 
+  const ariaLabelHint = (!ariaLabelledBy) ? `${removedPillInteractionHint}, ${ariaLabel}` : undefined;
+  const ariaAttrs = (React.Children.count(children)) ? {
+    'aria-live': 'assertive',
+    'aria-label': ariaLabelHint,
+    'aria-labelledby': ariaLabelledBy,
+    'aria-describedby': pillGroupAriaDescribedBy,
+    role: 'list',
+    tabIndex: containerTabindex,
+  } : { role: 'group' };
+
   return (
     <div
       {...customProps}
       {...filterPillsProps}
-      aria-live="assertive"
-      aria-label={!ariaLabelledBy ? `${removedPillInteractionHint} ${containerHint.current}` : undefined}
-      aria-labelledby={ariaLabelledBy}
-      aria-describedby={pillGroupAriaDescribedBy}
+      {...ariaAttrs}
       className={pillListClassNames}
       ref={filterPillsRef}
-      role="list"
-      tabIndex={containerTabindex}
     >
       <VisuallyHiddenText
         aria-live="polite"
         id={pillGroupInteractionHintID}
         text={pillGroupInteractionHint}
-        aria-hidden="true"
       />
       {children ? renderChildren(children) : []}
       {(isCollapsible && rollUpCount > 0) && (