+++ packages/terra-tabs/src/common-tabs/_Tab.jsx
@@ -192,8 +192,8 @@ const Tab = ({
       if (index === 0) {
         element = document.getElementById(tabIds[index + 1]);
       }
-      const ariaLabel = label ? `${label} ${deleteTabLabel}, ${element.title}` : '';
       if (element) {
+        const ariaLabel = label ? `${label} ${deleteTabLabel}, ${element.title}` : '';
         element.setAttribute('aria-label', ariaLabel);
         element.focus();
         element.addEventListener('blur', () => {