+++ packages/terra-folder-tree/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.0.0-alpha.3 - (December 11, 2023)
+
+* Changed
+  * Updated cursor style to pointer for clickable regions.
+
+## 1.0.0-alpha.2 - (December 5, 2023)
+
+* Added
+  * Added the ability to expand and collapse folders.
+
 ## 1.0.0-alpha.1 - (November 22, 2023)
 
 * Added