+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Fixed
+ * Added check to throw error when invalid data type is used for action buttons.
+
+## 4.37.0 - (December 11, 2023)
+
+* Changed
+  * Changed the notification dialog heading "Alert" from 'div' tag with an 'h2' element.
+
 ## 4.36.0 - (November 23, 2023)
 
 * Fixed